bin(dir_size): generalise `cargo_size` and switch to `pdu`
This adds a new script, `dir_size`, which does what `cargo_size`
currently does but allows taking an argument for a specific directory.
Because this makes the logic of `cargo_size` redundant, this patch also
removes the `cargo_size` script and adds an alias in its place.

Additionally, rather than using `du -sh` which pegs a single core, this
switches to using `pdu` to vastly improve performance on large
directories.