bin(mem_used): add script to get non-ZFS memory used
Because ZFS's cache (see `arc_size`) is implemented in userland on Linux
it causes `free` to become fairly useless for getting an idea of how
much memory is actually being taken by applications in the system. This
script simply subtracts the ARC's amount from the total usage and
returns a more realistic value.