git: make configs more generic for use by other machines
The global git config previously had hardcoded absolute paths. This
patch changes those to home based (`~`) paths and adds the the `EDITOR`
environmental variable to the shell profile which git will
respect.

Additionally, to account for more SSH keys in the future, a script was
added to the new `ssh` package which loops through the provided SSH keys
and adds them to `ssh-agent`. This list (and the git `user.signingKey`)
assumes that the names are consistent between *machines*, but the actual
keys themselves should **not** be for security reasons.

Keys that aren't used by on a machine will simply silently fail.